March 5, 2012 Occupy Rochester Hydrofracking Working Group
DRAFT STATEMENT FOR MARCH 16, 2012 NEW YORK STATE REPUBLICAN CONVENTION PROTEST IN ROCHESTER, NEW YORK

Governor Cuomo has recently stated that a decision on Hydrofracking could be announced within the next 2 months. The DEC's dSGEIS is fatally flawed. The DEC fails to study the cumulative impacts on our health from water contamination and air pollution from the shale gas extraction industry. The DEC fails to study the long term impacts of increased exposure of our children, to chemicals used in fracking that are known carcinogens and endocrine disruptors. The DEC fails to study the cumulative effects of these chemicals and radioactive frack fluids being sprayed as brine on our roads, and migrating into streams, rivers, and drinking water supplies. The DEC fails to study the long term effects on our farms and foodshed. 

All we need to do is to look to our neighbors in PA to see the effects of 4 years of fracking. There are many PA residents who have had their wells contaminated, and now must have water delivered for drinking, cooking, and bathing. There are reports of sickness and deaths both in humans, and in farm animals. Abnormally high levels of Barium, Strontium, and Arsenic have been found in water and in the blood of residents next to gas wells. We have read report after report of Well Blowouts, spills, leaks, explosions, and accidents, methane migration and chemical contamination of water sources, as well as other socioeconomic impacts, such as increased crime(drug use and rape) as the result of the influx of a transient work force.

The political fracking game is in play, with both the Democratic Governor and the Republican Senate poised to take credit for looking wise while studying the issue, and then ultimately giving in to their Corporate masters. The game will be to give pseudo protection to the NYC, Syracuse, Fingerlakes, and Cooperstown areas, while allowing Fracking to occur in The Southern Tier. The game is to fracture the state, and then Frack us, all while giving political cover to both corporate owned parties. But we know, due to the economics of the gas industry, that it is not currently profitable to frack. Gas is selling on the international market for 6 times as much as the domestic price. The game now, is to gain entry into NYS, build infrastructure of pipelines, compressor stations, and LNG export facilities, so that this shale gas will be sold to the highest bidder...more than likely, China.

There are a few bills in the NYS Senate that could stop the fracking nightmare that lies just south of the border in PA, from becoming a reality in NY.

The Hazardous Waste Bill,(S4616), already passed through the Assembly and now sits and waits for consideration by the NYS Senate. Currently, Frack Fluid wastes from PA are being trucked into NY state and have even been sprayed on our roads as "brine". The EPA has stated that the DEC has not accounted for the regulation and treatment of radioactivity in frack fluid, and yet currently it is unregulated and can be sprayed directly on our roads, leaching off into our farmlands, streams and rivers. There is no compelling reason why waste produced from oil and natural gas activities that meets the definition of hazardous waste, should not be subject to the same laws regarding generation, transportation. treatment, storage and disposal as other hazardous wastes. This Bill would change that by classifying this waste as hazardous and regulating it accordingly.

SB4220(Avella) Colton A7218A would Ban the hydraulic fracturing process in NYS. More and more local towns are taking it upon themselves to ban the destructive forces of the gas industry, by zoning and land use ordinances, or temporarily by moratoria bills. The NYS legislature needs to listen to the citizens of NYS who wish to protect their communities, their land, water, air, food, farms, and children's health.


But we have a problem. The Republican Leadership has not lead on this issue. The Republican leadership has blocked every bill that would serve to protect us, from ever coming to a vote on the full senate floor. State Senator Tom Libous' claims to be neutral and that he wants this decided by science and facts, not by emotions. He pretends to be neutral while advocating for supposed "Safe Drilling" and even has a website for this purpose. 

Republican Senator Ball knows the horrors of hydrofracking after touring Fracked areas of Pennsylvania and meeting with the victims of the un-natural shale gas industry. Senator Ball has stated that he would support legislation for a moratorium, IF the legislation were allowed to proceed to the Senate floor. But even Senator Ball could do so much more to co-sponsor Senator Avella's Ban bill. 

We call on the Republican Senate Leaders to quit blocking legislation that would protect the people of NYS. We call on the Republicans to actively step up and oppose Gov. Cuomo's RUSH to frack NYS. We call upon the Republican Senators to protect the environmental, economic, and public health interests of ALL of NYS. It is a crime to poison us.

~Harry Davis